State added with database reference and role model done

+++ backend/src/models/request_model.rs
@@ -1,13 +1,14 @@
 use serde::{Serialize, Deserialize};
 use crate::models::{request_type_model::RequestType, user_model::UserPublic};
+use sqlx::FromRow;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize,FromRow)]
 pub struct Request {
-    pub id: u64,
-    pub user: u64,
+    pub id: i32,
+    pub user: i64,
     pub request_type: RequestType,
     pub reason: Option<String>,
-    pub days: Vec<u64>,
+    pub days: Vec<i64>,
     pub status: Option<Status>,     //Every request should be created with pending
     pub created_at: Option<String>, //Its optional because it shouldnt be on payload
 }
