+++ backend/src/models/team_model.rs
@@ -1,10 +1,11 @@
 use serde::{Serialize, Deserialize};
+use sqlx::FromRow;
 
 use crate::models::user_model::UserPublic;
 
-#[derive(Serialize, Deserialize)] 
+#[derive(Serialize, Deserialize, FromRow)] 
 pub struct Team {
-    pub id: u64,
+    pub id: i32,
     pub team_name: String,
     pub desc: Option<String>,
     pub members: Option<Vec<UserPublic>>,
